Fixup usage of lex/yacc in Android.mk files
Add M4 environment variable for flex so it uses our prebuilt instead of
the one in $PATH.

Add dependencies on the prebuilt executables and data files for flex and
bison.

Use .KATI_IMPLICIT_OUTPUTS to specify that a single rule creates
multiple files.

src/mesa/program/... was standard enough to use the built-in lex
support, so move program_lexer.l to the sources list.

Switch from custom macros to using the common transform-y-to-c-or-cpp
macro the build system uses for yacc. We can't just add them to the
sources list because the filenames are different and the YACCFLAGS
differ in the same module.

Bug: 117561006
Test: mmm external/mesa3d
Change-Id: I615d85714409cb03221a2bebb0bdd5d25e7313d9

Signed-off-by: Amit Pundir <amit.pundir@linaro.org>